Adds the images in game notes
I copied what the yaml PR #8966 had for description. I think it should work but not 100% sure. @DanVanAtta so when you get time I think this is all that's needed and then have the #8966 fixed and it should be ready to roll. I know @merleyn is excited to share the fruits of his labor :)

Thanks